refactor: UI로직에 있던 getFormattedTime() 함수 -> 도메인 모델 함수로 이동
