+++ core/domain/src/main/java/com/example/domain/model/home/Post.kt
@@ -1,5 +1,6 @@
 package com.example.domain.model.home
 
+import java.time.Duration
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
@@ -13,7 +14,35 @@ data class PostFeed(
     val commentCount: Int = 0,
     val isVerified : Boolean = false,
     val imageUri: String = "",
-)
+) {
+     fun getFormattedTime(): String {
+        val now = LocalDateTime.now()
+        val duration = Duration.between(createdAt, now)
+
+        // 1일 이상일 경우
+        if (duration.toDays() >= 1) {
+            return if (duration.toDays() <= 7) {
+                "${duration.toDays()}일 전"
+            } else {
+                val formatter = DateTimeFormatter.ofPattern("M/d")
+                createdAt.format(formatter)
+            }
+        }
+
+        // 1시간 이상일 경우
+        if (duration.toHours() >= 1) {
+            return "${duration.toHours()}시간 전"
+        }
+
+        // 1분 이상일 경우
+        if (duration.toMinutes() >= 1) {
+            return "${duration.toMinutes()}분 전"
+        }
+
+        // 1분 미만일 경우
+        return "방금 전"
+    }
+}
 
 data class PostDetail(
     val postType : PostType,
